Refactor + fix some hive tests
